Adding TrashClient
Code for TrashClient added.
